EnglishProject/src/main/python/requirements.txt
google-generativeai>=0.5.0
mysql-connector-python>=8.0.0
edge-tts